[C] Move routes folder
