+++ components/composed/community/CommunityLayout/CommunityLayout.tsx
@@ -4,7 +4,7 @@ import { CommunityLayoutQuery } from "@/relay/CommunityLayoutQuery.graphql";
 import useAuthenticatedQuery from "hooks/useAuthenticatedQuery";
 import { ModelHeader } from "components/composed/model";
 import { useRouter } from "next/router";
-import { routeQueryArrayToString } from "helpers";
+import { routeQueryArrayToString } from "routes";
 
 export default function CommunityLayout({ children }) {
   const router = useRouter();
