+++ helpers/index.ts
@@ -1,4 +1,3 @@
 export { default as encodeSvg } from "./encodeSvg";
 export * from "./dates";
-export * from "./routes";
 export * from "./arrays";
